requirements.txt
asgiref==3.11.0
Django==5.2.8
django-filter==25.2
django-unfold==0.77.1
djangorestframework==3.16.1
pillow==12.1.0
sqlparse==0.5.5
